Fix querying on (Generic)EmbeddedDocument subclasses fields
